requirements.txt
Flask
gunicorn
python-dotenv
google-generativeai
psycopg2-binary